requirements.txt
fastapi>=0.100.0
uvicorn==0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
SQLAlchemy==1.4.27
apscheduler==3.9.1
redis==4.6.0
python-dotenv==1.0.1
